+++ src/components/capabilities/implementation/fission-lobby.ts
@@ -3,7 +3,6 @@ import * as Uint8arrays from "uint8arrays"
 import * as Base64 from "../../../common/base64.js"
 import * as Capabilities from "../../../capabilities.js"
 import * as Crypto from "../../../components/crypto/implementation.js"
-import * as Depot from "../../../components/depot/implementation.js"
 import * as DID from "../../../did/index.js"
 import * as Fission from "../../../common/fission.js"
 import * as Path from "../../../path/index.js"
@@ -13,7 +12,6 @@ import * as Ucan from "../../../ucan/index.js"
 import { Implementation, RequestOptions } from "../implementation.js"
 import { Maybe } from "../../../common/types.js"
 import { VERSION } from "../../../common/version.js"
-import { decodeCID } from "../../../common/cid.js"
 
 
 // 🧩
@@ -21,7 +19,6 @@ import { decodeCID } from "../../../common/cid.js"
 
 export type Dependencies = {
   crypto: Crypto.Implementation
-  depot: Depot.Implementation
 }
 
 
@@ -34,30 +31,19 @@ export async function collect(
   dependencies: Dependencies
 ): Promise<Maybe<Capabilities.Capabilities>> {
   const url = new URL(self.location.href)
-  const authorised = url.searchParams.get("authorised")
-  if (!authorised) return null
-
   const username = url.searchParams.get("username") ?? ""
-  const secrets = await retry(
-    async () => translateClassifiedInfo(
-      dependencies,
-      authorised === "via-postmessage"
-        ? await getClassifiedViaPostMessage(endpoints, dependencies.crypto)
-        : JSON.parse(
-          Uint8arrays.toString(
-            await dependencies.depot.getUnixFile(
-              decodeCID(authorised)
-            ),
-            "utf8"
-          )
-        )
-    ),
+  if (!username) return null
+
+  const info = await retry(
+    () => getClassifiedViaPostMessage(endpoints, dependencies.crypto),
     {
       tries: 20,
       timeout: 60000,
       timeoutMessage: "Trying to retrieve UCAN(s) and readKey(s) from the auth lobby timed out after 60 seconds."
     }
   )
+
+  const secrets = await translateClassifiedInfo(dependencies, info)
 
   if (!secrets) {
     throw new Error("Failed to retrieve secrets from lobby url parameters")
@@ -99,26 +85,26 @@ export async function request(
 
   // Compile params
   const params = [
-    [ "didExchange", exchangeDid ],
-    [ "didWrite", writeDid ],
-    [ "redirectTo", redirectTo ],
-    [ "sdk", VERSION.toString() ],
-    [ "sharedRepo", sharedRepo ? "t" : "f" ],
-    [ "sharing", sharing ? "t" : "f" ]
+    ["didExchange", exchangeDid],
+    ["didWrite", writeDid],
+    ["redirectTo", redirectTo],
+    ["sdk", VERSION.toString()],
+    ["sharedRepo", sharedRepo ? "t" : "f"],
+    ["sharing", sharing ? "t" : "f"]
 
   ].concat(
-    app ? [ [ "appFolder", `${app.creator}/${app.name}` ] ] : [],
-    fs?.private ? fs.private.map(p => [ "privatePath", Path.toPosix(p, { absolute: true }) ]) : [],
-    fs?.public ? fs.public.map(p => [ "publicPath", Path.toPosix(p, { absolute: true }) ]) : [],
-    raw ? [ [ "raw", Base64.urlEncode(JSON.stringify(raw)) ] ] : [],
+    app ? [["appFolder", `${app.creator}/${app.name}`]] : [],
+    fs?.private ? fs.private.map(p => ["privatePath", Path.toPosix(p, { absolute: true })]) : [],
+    fs?.public ? fs.public.map(p => ["publicPath", Path.toPosix(p, { absolute: true })]) : [],
+    raw ? [["raw", Base64.urlEncode(JSON.stringify(raw))]] : [],
     options.extraParams ? Object.entries(options.extraParams) : []
 
   ).concat((() => {
     const apps = platform?.apps
 
     switch (typeof apps) {
-      case "string": return [ [ "app", apps ] ]
-      case "object": return apps.map(a => [ "app", a ])
+      case "string": return [["app", apps]]
+      case "object": return apps.map(a => ["app", a])
       default: return []
     }
 
@@ -127,7 +113,7 @@ export async function request(
   // And, go!
   window.location.href = endpoints.lobby + "?" +
     params
-      .map(([ k, v ]) => encodeURIComponent(k) + "=" + encodeURIComponent(v))
+      .map(([k, v]) => encodeURIComponent(k) + "=" + encodeURIComponent(v))
       .join("&")
 }
 
@@ -240,7 +226,7 @@ async function translateClassifiedInfo(
   // This easy way of detection works because the decrypted session key is encoded in base 64.
   // That means it'll only ever use the first byte to encode it, and if it were UTF-16 it would
   // split up the two bytes. Hence we check for the second byte here.
-  const isUtf16 = rawSessionKey[ 1 ] === 0
+  const isUtf16 = rawSessionKey[1] === 0
 
   const sessionKey = isUtf16
     ? Uint8arrays.fromString(
@@ -267,7 +253,7 @@ async function translateClassifiedInfo(
     isLobbySecrets(secrets)
       ? Object
         .entries(secrets.fs)
-        .map(([ posixPath, { bareNameFilter, key } ]) => {
+        .map(([posixPath, { bareNameFilter, key }]) => {
           return {
             bareNameFilter: bareNameFilter,
             path: Path.fromPosix(posixPath),
@@ -295,20 +281,23 @@ async function retry<T>(
   action: () => Promise<T>,
   options: { tries: number; timeout: number; timeoutMessage: string }
 ): Promise<T> {
-  return await Promise.race([
-    action(),
-    new Promise<T>((resolve, reject) => {
-      if (options.tries > 0) return setTimeout(
-        () => retry(action, { ...options, tries: options.tries - 1 }).then(resolve, reject),
-        options.timeout
-      )
+  return new Promise((resolve, reject) => {
+    if (options.tries > 0) {
+      const unoMas = () => {
+        retry(action, { ...options, tries: options.tries - 1 })
+      }
 
-      return setTimeout(
-        () => reject(new Error(options.timeoutMessage)),
-        options.timeout
-      )
-    })
-  ])
+      const timeoutId = setTimeout(unoMas, options.timeout)
+
+      action()
+        .then(resolve, unoMas)
+        .finally(() => clearTimeout(timeoutId))
+
+    } else {
+      reject(new Error(options.timeoutMessage))
+
+    }
+  })
 }